@@ -2,10 +2,13 @@
   "name": "shadowform-monorepo",
   "description": "Shadowform monorepo",
   "dependencies": {
-    "mobx": "^4.1.1",
+    "mobx": "^6.0.1",
     "mobx-react": "^5.1.2",
-    "react": "^16.4.1",
-    "react-dom": "^16.4.1"
+    "mobx-react-lite": "^3.0.0",
+    "raw-loader": "^4.0.2",
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1",
+    "style-loader": "^2.0.0"
   },
   "devDependencies": {
     "babel-eslint": "^9.0.0",
@@ -19,9 +22,9 @@
     "eslint-plugin-jsx-a11y": "^6.1.1",
     "eslint-plugin-prettier": "^2.6.0",
     "eslint-plugin-react": "^7.11.0",
-    "gnoll": "^0.6.0-7",
+    "gnoll": "^0.7.0-alpha.4",
     "gnoll-styles": "^0.6.0-1",
-    "lerna": "^3.2.1",
+    "lerna": "^3.4.0",
     "prettier": "^1.11.1"
   },
   "author": "sunflowerdeath",
