@@ -32,8 +32,8 @@
     "prop-types": "^15.6.1"
   },
   "peerDependencies": {
-    "mobx": "^4.1.1",
-    "mobx-react": "^5.1.2",
-    "react": "^16.4.1"
+    "mobx": "^6.0.0",
+    "mobx-react": "^3.0.0",
+    "react": "^16.8.0"
   }
 }
