@@ -1,5 +1,5 @@
-import React, { Component } from 'react'
-import { observer } from 'mobx-react'
+import React from 'react'
+import { observer, useLocalObservable } from 'mobx-react-lite'
 import { FormStore } from 'shadowform'
 
 import Field from './Field'
@@ -41,40 +41,26 @@ const submit = async form => {
 	alert(`Submitted:\n${JSON.stringify(form.values)}`)
 }
 
-@observer
-class SubmitValidationExample extends Component {
-	constructor() {
-		super()
-		this.form = createForm()
-	}
-
-	render() {
-		const form = this.form
-		return (
-			<div>
-				<label style={{ marginBottom: 5 }}>Username:</label>
-				<Field
-					field={form.fields.username}
-					showRequiredError="onBlurTouched"
-				/>
-				<label style={{ marginBottom: 5 }}>Password:</label>
-				<Field
-					field={form.fields.password}
-					showRequiredError="onBlurTouched"
-				/>
-				<button
-					className="button"
-					disabled={!form.isValid}
-					onClick={() => submit(this.form)}
-				>
-					Submit
-				</button>
-				<button className="button" onClick={() => form.reset()}>
-					Reset
-				</button>
-			</div>
-		)
-	}
+const SubmitValidationExample = () => {
+	const form = useLocalObservable(createForm)
+	return (
+		<div>
+			<label style={{ marginBottom: 5 }}>Username:</label>
+			<Field field={form.fields.username} showRequiredError="onBlurTouched" />
+			<label style={{ marginBottom: 5 }}>Password:</label>
+			<Field field={form.fields.password} showRequiredError="onBlurTouched" />
+			<button
+				className="button"
+				disabled={!form.isValid}
+				onClick={() => submit(form)}
+			>
+				Submit
+			</button>
+			<button className="button" onClick={() => form.reset()}>
+				Reset
+			</button>
+		</div>
+	)
 }
 
-export default SubmitValidationExample
+export default observer(SubmitValidationExample)
